Add: ランキングの内容を別コンポーネントに移した
